Suggest using localhost on windows
For my test 0.0.0.0 did not work on Windows, but localhost did.  I used the pixi method to set up the server.

It looks like 0.0.0.0 is suggested automatically by the server.  It could make sense to override that or add a message suggesting the localhost version of the URL for configuration.